refactor: more changes to modules and imports
