+++ router/src/ingress/mod.rs
@@ -1,2 +1,2 @@
-pub mod server;
+pub mod http_server;
 pub mod types;
